Move most cdk to the same directory
