+++ source/lib/appregistry/applyAppRegistry.ts
