+++ source/lib/member/redshift-audit-logging.ts
