+++ source/test/orchestrator_logs.test.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import * as cdk from 'aws-cdk-lib';
-import { OrchLogStack } from '../solution_deploy/lib/orchestrator-log-stack';
+import { OrchLogStack } from '../lib/orchestrator-log-stack';
 import { AwsSolutionsChecks } from 'cdk-nag';
 import { Aspects, DefaultStackSynthesizer } from 'aws-cdk-lib';
 import { Template } from 'aws-cdk-lib/assertions';
